requirements.txt
selenium==4.19.0
lxml==5.2.1
beautifulsoup4==4.12.3
python-dotenv==1.0.0
crewai
crewai-tools
langchain-core
langchain-community
groq
langchain-groq
agno
duckduckgo-search